+++ app/src/main/java/com/electropeyk/squenda/activities/VideoListActivity.kt
@@ -4,14 +4,14 @@ package com.electropeyk.squenda.activities
 import android.content.Intent
 import android.net.Uri
 import android.os.Bundle
+import android.os.Handler
 import android.view.View
+import android.view.animation.AnimationUtils
 import android.widget.Toast
 import androidx.appcompat.app.AppCompatActivity
 import androidx.core.content.ContextCompat
 import androidx.recyclerview.widget.GridLayoutManager
 import androidx.recyclerview.widget.RecyclerView
-import com.crowdfire.cfalertdialog.CFAlertDialog
-import com.electropeyk.squenda.R
 import com.electropeyk.squenda.adpter.VideoRecyclerViewAdapter
 import com.electropeyk.squenda.utils.Common
 import com.electropeyk.squenda.utils.Common.ABSOLUTE_PATH_NAMES_VIDEO_LIST
@@ -24,11 +24,13 @@ import java.text.SimpleDateFormat
 import java.util.*
 import kotlin.collections.ArrayList
 
+
 class VideoListActivity : AppCompatActivity(), VideoRecyclerViewAdapter.ItemClickListener,
     VideoRecyclerViewAdapter.ItemLongClickListener {
     lateinit var recyclerView: RecyclerView
     lateinit var adapter: VideoRecyclerViewAdapter
     private val NUM_COLUMNS = 6
+    private var isOpen: Boolean = false
 
     private var mDividerItemDecoration: RecyclerView.ItemDecoration? = null
     override fun onCreate(savedInstanceState: Bundle?) {
@@ -37,7 +39,8 @@ class VideoListActivity : AppCompatActivity(), VideoRecyclerViewAdapter.ItemClic
         setContentView(com.electropeyk.squenda.R.layout.activity_video_list)
         overridePendingTransition(com.electropeyk.squenda.R.anim.fade_in, com.electropeyk.squenda.R.anim.fade_out)
         Paper.init(this)
-        recyclerView = findViewById(R.id.recycle_videos)
+        fr_msg.setVisibility(View.GONE)
+        recyclerView = findViewById(com.electropeyk.squenda.R.id.recycle_videos)
         ABSOLUTE_PATH_NAMES_VIDEO_LIST = Paper.book(Common.DATABASE).read(Common.ABSOLUTE_PATH_NAMES_VIDEO)
         initRecyclerView()
         btn_share.setVisibility(View.INVISIBLE)
@@ -47,13 +50,13 @@ class VideoListActivity : AppCompatActivity(), VideoRecyclerViewAdapter.ItemClic
             val intent = Intent(this, MyHomeActivity::class.java)
             // start your next activity
             startActivity(intent)
-            overridePendingTransition(R.anim.fade_in, R.anim.fade_out)
+            overridePendingTransition(com.electropeyk.squenda.R.anim.fade_in, com.electropeyk.squenda.R.anim.fade_out)
             finish()
         }
         btn_setting_video_list.setOnClickListener {
             val intent = Intent(this, SettingActivity::class.java)
             // start your next activity
-            overridePendingTransition(R.anim.fade_in, R.anim.fade_out)
+            overridePendingTransition(com.electropeyk.squenda.R.anim.fade_in, com.electropeyk.squenda.R.anim.fade_out)
             startActivity(intent)
             finish()
         }
@@ -62,36 +65,43 @@ class VideoListActivity : AppCompatActivity(), VideoRecyclerViewAdapter.ItemClic
             val intent = Intent(this, MediaActivity::class.java)
             // start your next activity
             startActivity(intent)
-            overridePendingTransition(R.anim.fade_in, R.anim.fade_out)
+            overridePendingTransition(com.electropeyk.squenda.R.anim.fade_in, com.electropeyk.squenda.R.anim.fade_out)
             finish()
         }
 
         btn_trash.setOnClickListener {
             if (VIDEO_NUM_SELECCTED.size > 0) {
+                fr_msg.setVisibility(View.VISIBLE)
+                val RightSwipe = AnimationUtils.loadAnimation(this, com.electropeyk.squenda.R.anim.right_swipe)
+                fr_msg.startAnimation(RightSwipe)
+                isOpen = true
+                val handler = Handler()
+                handler.postDelayed(Runnable {
+                    if (isOpen) {
+                        val RightSwipe = AnimationUtils.loadAnimation(this, com.electropeyk.squenda.R.anim.left_swipe)
+                        fr_msg.startAnimation(RightSwipe)
+                        fr_msg.setVisibility(View.GONE)
+                    }
 
-                val builder = CFAlertDialog.Builder(this)
-                    .setDialogStyle(CFAlertDialog.CFAlertStyle.NOTIFICATION)
-                    .setTitle("Delete Videos")
-                    .setMessage("Are you sure to delete selected videos?")
-                    .addButton(
-                        "Yes", -1, -1,
-                        CFAlertDialog.CFAlertActionStyle.DEFAULT, CFAlertDialog.CFAlertActionAlignment.JUSTIFIED
-                    ) { dialog, which ->
-                        deleteVideos()
-                        dialog.dismiss()
-                    }.addButton(
-                        "NO", -1, -1,
-                        CFAlertDialog.CFAlertActionStyle.NEGATIVE, CFAlertDialog.CFAlertActionAlignment.JUSTIFIED
-                    ) { dialog, which -> dialog.dismiss() }
+                }, 4000)
 
-                // Show the alert
-                builder.show()
+
             } else {
                 Toast.makeText(this, "There is not any item selected for delete", Toast.LENGTH_LONG).show()
             }
 
 
         }
+
+        btn_check.setOnClickListener {
+            deleteVideos()
+            val RightSwipe = AnimationUtils.loadAnimation(this, com.electropeyk.squenda.R.anim.left_swipe)
+            fr_msg.startAnimation(RightSwipe)
+            fr_msg.setVisibility(View.GONE)
+            isOpen=false
+
+        }
+
         btn_share.setOnClickListener {
             if (VIDEO_NUM_SELECCTED.size > 0) {
                 val imageUris = ArrayList<Uri>()
@@ -109,14 +119,14 @@ class VideoListActivity : AppCompatActivity(), VideoRecyclerViewAdapter.ItemClic
 
         }
 
-        img_back_video_list.setOnClickListener{
+        img_back_video_list.setOnClickListener {
             val intent = Intent(this, GalleryActivity::class.java)
             // start your next activity
             startActivity(intent)
-            overridePendingTransition(R.anim.fade_in, R.anim.fade_out)
+            overridePendingTransition(com.electropeyk.squenda.R.anim.fade_in, com.electropeyk.squenda.R.anim.fade_out)
             finish()
         }
-        txt_time_video_list.text= SimpleDateFormat("HH:mm", Locale.US).format( Date())
+        txt_time_video_list.text = SimpleDateFormat("HH:mm", Locale.US).format(Date())
 
         val thread = object : Thread() {
 
@@ -125,7 +135,7 @@ class VideoListActivity : AppCompatActivity(), VideoRecyclerViewAdapter.ItemClic
                     while (!this.isInterrupted) {
                         sleep(1000)
                         runOnUiThread {
-                            txt_time_video_list.text= SimpleDateFormat("HH:mm", Locale.US).format( Date())
+                            txt_time_video_list.text = SimpleDateFormat("HH:mm", Locale.US).format(Date())
                         }
                     }
                 } catch (e: InterruptedException) {
@@ -135,28 +145,32 @@ class VideoListActivity : AppCompatActivity(), VideoRecyclerViewAdapter.ItemClic
         }
 
         thread.start()
-        var dayOfMonth=Calendar.getInstance().get(Calendar.DAY_OF_MONTH)
+        var dayOfMonth = Calendar.getInstance().get(Calendar.DAY_OF_MONTH)
         val day = Common.days[Calendar.getInstance().get(Calendar.DAY_OF_WEEK) - 1]
         val month = Common.months[Calendar.getInstance().get(Calendar.MONTH)]
-        txt_date_video_list.text= "$day,$month $dayOfMonth"
+        txt_date_video_list.text = "$day,$month $dayOfMonth"
 
 
     }
 
     private fun deleteVideos() {
-        for (position in VIDEO_NUM_SELECCTED) {
+        if(ABSOLUTE_PATH_NAMES_VIDEO_LIST.size==VIDEO_NUM_SELECCTED.size)
+            ABSOLUTE_PATH_NAMES_VIDEO_LIST=ArrayList()
+        else {
+            for (position in VIDEO_NUM_SELECCTED) {
 
-            val fdelete = File(ABSOLUTE_PATH_NAMES_VIDEO_LIST[position])
-            if (fdelete.exists()) {
-                if (fdelete.delete()) {
-                    ABSOLUTE_PATH_NAMES_VIDEO_LIST.removeAt(position)
+                val fdelete = File(ABSOLUTE_PATH_NAMES_VIDEO_LIST[position])
+                if (fdelete.exists()) {
+                    if (fdelete.delete()) {
+                        ABSOLUTE_PATH_NAMES_VIDEO_LIST.removeAt(position)
+                    }
                 }
-                }
 
 
+            }
         }
         adapter = VideoRecyclerViewAdapter(this, ABSOLUTE_PATH_NAMES_VIDEO_LIST)
-        recyclerView.adapter?.notifyDataSetChanged()
+        recyclerView!!.adapter = adapter
         Paper.book(Common.DATABASE).write(Common.ABSOLUTE_PATH_NAMES_VIDEO, Common.ABSOLUTE_PATH_NAMES_VIDEO_LIST)
     }
